+++ zetaclient/bitcoin/inbound_tracker.go
@@ -10,10 +10,11 @@ import (
 	"github.com/zeta-chain/zetacore/zetaclient/zetabridge"
 )
 
-func (ob *BTCChainClient) ExternalChainWatcherForNewInboundTrackerSuggestions() {
-	ticker, err := types.NewDynamicTicker("Bitcoin_WatchInTx_InboundTrackerSuggestions", ob.GetChainParams().InTxTicker)
+// WatchIntxTracker watches zetacore for bitcoin intx trackers
+func (ob *BTCChainClient) WatchIntxTracker() {
+	ticker, err := types.NewDynamicTicker("Bitcoin_WatchIntxTracker", ob.GetChainParams().InTxTicker)
 	if err != nil {
-		ob.logger.WatchInTx.Err(err).Msg("error creating ticker")
+		ob.logger.InTx.Err(err).Msg("error creating ticker")
 		return
 	}
 
@@ -24,13 +25,16 @@ func (ob *BTCChainClient) ExternalChainWatcherForNewInboundTrackerSuggestions() 
 			if flags := ob.coreContext.GetCrossChainFlags(); !flags.IsInboundEnabled {
 				continue
 			}
+			if !ob.GetChainParams().IsSupported {
+				continue
+			}
 			err := ob.ObserveTrackerSuggestions()
 			if err != nil {
-				ob.logger.WatchInTx.Error().Err(err).Msg("error observing in tx")
+				ob.logger.InTx.Error().Err(err).Msgf("error observing intx tracker for chain %d", ob.chain.ChainId)
 			}
-			ticker.UpdateInterval(ob.GetChainParams().InTxTicker, ob.logger.WatchInTx)
+			ticker.UpdateInterval(ob.GetChainParams().InTxTicker, ob.logger.InTx)
 		case <-ob.stop:
-			ob.logger.WatchInTx.Info().Msg("ExternalChainWatcher for BTC inboundTrackerSuggestions stopped")
+			ob.logger.InTx.Info().Msgf("WatchIntxTracker stopped for chain %d", ob.chain.ChainId)
 			return
 		}
 	}
@@ -42,12 +46,12 @@ func (ob *BTCChainClient) ObserveTrackerSuggestions() error {
 		return err
 	}
 	for _, tracker := range trackers {
-		ob.logger.WatchInTx.Info().Msgf("checking tracker with hash :%s and coin-type :%s ", tracker.TxHash, tracker.CoinType)
+		ob.logger.InTx.Info().Msgf("checking tracker with hash :%s and coin-type :%s ", tracker.TxHash, tracker.CoinType)
 		ballotIdentifier, err := ob.CheckReceiptForBtcTxHash(tracker.TxHash, true)
 		if err != nil {
 			return err
 		}
-		ob.logger.WatchInTx.Info().Msgf("Vote submitted for inbound Tracker,Chain : %s,Ballot Identifier : %s, coin-type %s", ob.chain.ChainName, ballotIdentifier, coin.CoinType_Gas.String())
+		ob.logger.InTx.Info().Msgf("Vote submitted for inbound Tracker,Chain : %s,Ballot Identifier : %s, coin-type %s", ob.chain.ChainName, ballotIdentifier, coin.CoinType_Gas.String())
 	}
 	return nil
 }
@@ -72,13 +76,13 @@ func (ob *BTCChainClient) CheckReceiptForBtcTxHash(txHash string, vote bool) (st
 	if len(blockVb.Tx) <= 1 {
 		return "", fmt.Errorf("block %d has no transactions", blockVb.Height)
 	}
-	depositorFee := CalcDepositorFee(blockVb, ob.chain.ChainId, ob.netParams, ob.logger.WatchInTx)
+	depositorFee := CalcDepositorFee(blockVb, ob.chain.ChainId, ob.netParams, ob.logger.InTx)
 	tss, err := ob.zetaClient.GetBtcTssAddress(ob.chain.ChainId)
 	if err != nil {
 		return "", err
 	}
 	// #nosec G701 always positive
-	event, err := GetBtcEvent(*tx, tss, uint64(blockVb.Height), &ob.logger.WatchInTx, ob.netParams, depositorFee)
+	event, err := GetBtcEvent(*tx, tss, uint64(blockVb.Height), &ob.logger.InTx, ob.netParams, depositorFee)
 	if err != nil {
 		return "", err
 	}
@@ -94,10 +98,10 @@ func (ob *BTCChainClient) CheckReceiptForBtcTxHash(txHash string, vote bool) (st
 	}
 	zetaHash, ballot, err := ob.zetaClient.PostVoteInbound(zetabridge.PostVoteInboundGasLimit, zetabridge.PostVoteInboundExecutionGasLimit, msg)
 	if err != nil {
-		ob.logger.WatchInTx.Error().Err(err).Msg("error posting to zeta core")
+		ob.logger.InTx.Error().Err(err).Msg("error posting to zeta core")
 		return "", err
 	} else if zetaHash != "" {
-		ob.logger.WatchInTx.Info().Msgf("BTC deposit detected and reported: PostVoteInbound zeta tx hash: %s inTx %s ballot %s fee %v",
+		ob.logger.InTx.Info().Msgf("BTC deposit detected and reported: PostVoteInbound zeta tx hash: %s inTx %s ballot %s fee %v",
 			zetaHash, txHash, ballot, depositorFee)
 	}
 	return msg.Digest(), nil